extract flter panel utils
